requirements.txt
arabic-reshaper==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
filelock==3.16.1
fsspec==2024.12.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
Pillow==9.5.0
python-bidi==0.6.3
requests==2.31.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
urllib3==2.3.0
huggingface_hub
python-dotenv==1.2.1
datasets
pytest
tqdm
PyDrive2
google-auth
google-auth-oauthlib
google-auth-httplib2